move some classes in order to reduce public footprint
