+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/HbefaRoadTypeMapping.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.emissions.roadTypeMapping;
+package org.matsim.contrib.emissions;
 
 import com.google.inject.ProvidedBy;
 import org.matsim.api.core.v01.network.Link;
@@ -8,7 +8,7 @@ import org.matsim.contrib.emissions.utils.EmissionUtils;
 /**
  * Created by molloyj on 01.12.2017.
  */
-public abstract class HbefaRoadTypeMapping {
+abstract class HbefaRoadTypeMapping {
 
     public void addHbefaMappings(Network network) {
         for (Link link : network.getLinks().values()) {
