+++ contribs/emissions/src/main/java/org/matsim/contrib/emissions/LinkHbefaMapping.java
@@ -1,9 +1,9 @@
-package org.matsim.contrib.emissions.roadTypeMapping;
+package org.matsim.contrib.emissions;
 
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.contrib.emissions.utils.EmissionUtils;
 
-public class LinkHbefaMapping extends HbefaRoadTypeMapping {
+class LinkHbefaMapping extends HbefaRoadTypeMapping {
     @Override
     public String determineHebfaType(Link link) {
         return EmissionUtils.getHbefaRoadType(link);
